# services/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_workived_services_pkg_apperr.AppError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_workived_services_pkg_paginate.Meta:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
...
  internal_employee.CreateEmployeeRequest:
    properties:
      department_id:
        type: string
      email:
        maxLength: 255
        type: string
      employee_code:
        maxLength: 50
        type: string
      employment_type:
        enum:
        - full_time
        - part_time
        - contract
        - intern
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
      job_title:
        maxLength: 150
        type: string
      phone:
        maxLength: 30
        type: string
      start_date:
        type: string
    required:
    - email
    - employment_type
    - full_name
    - start_date
    type: object
  internal_employee.Employee:
    properties:
      base_salary:
        type: integer
      created_at:
        type: string
      custom_fields:
        items:
          type: integer
        type: array
      department_id:
        type: string
      email:
        type: string
      employee_code:
        type: string
      employment_type:
        type: string
      end_date:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      job_title:
        type: string
      organisation_id:
        type: string
      phone:
        type: string
      salary_currency:
        type: string
      start_date:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_employee.UpdateEmployeeRequest:
    properties:
      department_id:
        type: string
      employment_type:
        enum:
        - full_time
        - part_time
        - contract
        - intern
        type: string
      end_date:
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
      job_title:
        maxLength: 150
        type: string
      phone:
        maxLength: 30
        type: string
      status:
        enum:
        - active
        - on_leave
        - probation
        - inactive
        type: string
    type: object
  internal_organisation.CreateOrgRequest:
    properties:
      country_code:
        type: string
      currency_code:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
      timezone:
        maxLength: 50
        type: string
    required:
    - country_code
    - currency_code
    - name
    - slug
    - timezone
    type: object
  internal_organisation.InviteMemberRequest:
    properties:
      email:
        maxLength: 255
        type: string
      role:
        enum:
        - admin
        - member
        type: string
    required:
    - email
    - role
    type: object
  internal_organisation.Organisation:
    properties:
      country_code:
        type: string
      created_at:
        type: string
      currency_code:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      logo_url:
        type: string
      name:
        type: string
      plan:
        type: string
      plan_employee_limit:
        type: integer
      slug:
        type: string
      timezone:
        type: string
      work_days:
        items:
          type: integer
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: HR and operations superapp for SMB founders in Indonesia and UAE.
  title: Workived API
  version: "1.0"
paths:
...
  /employees:
    get:
      parameters:
      - description: Pagination cursor from previous response
        in: query
        name: cursor
        type: string
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: 'Filter: active | on_leave | probation | inactive'
        in: query
        name: status
        type: string
      - description: Filter by department UUID
        in: query
        name: department_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_employee.Employee'
                type: array
              meta:
                $ref: '#/definitions/github_com_workived_services_pkg_paginate.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: List employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_employee.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_employee.Employee'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "402":
          description: Free plan limit reached
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Create a new employee
      tags:
      - Employees
  /employees/{id}:
    delete:
      parameters:
      - description: Employee UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate (soft-delete) an employee
      tags:
      - Employees
    get:
      parameters:
      - description: Employee UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_employee.Employee'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Get a single employee by ID
      tags:
      - Employees
    put:
      consumes:
      - application/json
      parameters:
      - description: Employee UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_employee.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_employee.Employee'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - Employees
  /organisations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organisation details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_organisation.CreateOrgRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_organisation.Organisation'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "409":
          description: Slug already taken
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Create a new organisation
      tags:
      - Organisations
  /organisations/invitations:
    post:
      consumes:
      - application/json
      description: Requires owner or admin role.
      parameters:
      - description: Invitee email and role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_organisation.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "403":
          description: Insufficient role
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Invite a member to the organisation
      tags:
      - Organisations
  /organisations/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_organisation.Organisation'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Get the authenticated user's organisation
      tags:
      - Organisations
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"